BPAAS-1054: Move load function to api
