+++ cmd/api.go
@@ -2,15 +2,24 @@ package cmd
 
 import (
 	"context"
+	"fmt"
 	"net/http"
+	"net/url"
 	"time"
 
+	"github.com/bitcoin-sv/arc/api"
 	"github.com/bitcoin-sv/arc/api/handler"
+	"github.com/bitcoin-sv/arc/api/transactionHandler"
+	"github.com/bitcoin-sv/arc/blocktx"
+	"github.com/bitcoin-sv/arc/blocktx/blocktx_api"
+	"github.com/bitcoin-sv/arc/config"
 	"github.com/labstack/echo/v4"
 	echomiddleware "github.com/labstack/echo/v4/middleware"
 	apmecho "github.com/opentracing-contrib/echo"
 	"github.com/opentracing/opentracing-go"
+	"github.com/ordishs/go-bitcoin"
 	"github.com/ordishs/go-utils"
+	"github.com/pkg/errors"
 	"github.com/spf13/viper"
 )
 
@@ -40,7 +49,7 @@ func StartAPIServer(logger utils.Logger) (func(), error) {
 
 	// load the ARC handler from config
 	// If you want to customize this for your own server, see examples dir
-	if err := handler.LoadArcHandler(e, logger); err != nil {
+	if err := LoadArcHandler(e, logger); err != nil {
 		panic(err)
 	}
 
@@ -70,4 +79,105 @@ func StartAPIServer(logger utils.Logger) (func(), error) {
 			logger.Errorf("Error closing API echo server: %v", err)
 		}
 	}, nil
+}
+
+func LoadArcHandler(e *echo.Echo, logger utils.Logger) error {
+
+	// check the swagger definition against our requests
+	handler.CheckSwagger(e)
+
+	// Check the security requirements
+	//CheckSecurity(e, appConfig)
+
+	addresses := viper.GetString("metamorph.dialAddr")
+	if addresses == "" {
+		return fmt.Errorf("metamorph.dialAddr not found in config")
+	}
+
+	blocktxAddress := viper.GetString("blocktx.dialAddr")
+	if blocktxAddress == "" {
+		return fmt.Errorf("blocktx.dialAddr not found in config")
+	}
+
+	blockTxLogger, err := config.NewLogger()
+	if err != nil {
+		return fmt.Errorf("failed to create new logger: %v", err)
+	}
+
+	conn, err := blocktx.DialGRPC(blocktxAddress)
+	if err != nil {
+		return fmt.Errorf("failed to connect to block-tx server: %v", err)
+	}
+
+	bTx := blocktx.NewClient(blocktx_api.NewBlockTxAPIClient(conn), blocktx.WithLogger(blockTxLogger))
+
+	grpcMessageSize := viper.GetInt("grpcMessageSize")
+	if grpcMessageSize == 0 {
+		return fmt.Errorf("grpcMessageSize not found in config")
+	}
+
+	txHandler, err := transactionHandler.NewMetamorph(addresses, bTx, grpcMessageSize, logger)
+	if err != nil {
+		return err
+	}
+
+	var policy *bitcoin.Settings
+	policy, err = getPolicyFromNode()
+	if err != nil {
+		policy, err = handler.GetDefaultPolicy()
+		if err != nil {
+			return err
+		}
+	}
+
+	// TODO WithSecurityConfig(appConfig.Security)
+	apiHandler, err := handler.NewDefault(logger, txHandler, policy)
+	if err != nil {
+		return err
+	}
+
+	// Register the ARC API
+	api.RegisterHandlers(e, apiHandler)
+
+	return nil
+}
+
+func getPolicyFromNode() (*bitcoin.Settings, error) {
+	peerRpcPassword := viper.GetString("peerRpc.password")
+	if peerRpcPassword == "" {
+		return nil, errors.Errorf("setting peerRpc.password not found")
+	}
+
+	peerRpcUser := viper.GetString("peerRpc.user")
+	if peerRpcUser == "" {
+		return nil, errors.Errorf("setting peerRpc.user not found")
+	}
+
+	peerRpcHost := viper.GetString("peerRpc.host")
+	if peerRpcHost == "" {
+		return nil, errors.Errorf("setting peerRpc.host not found")
+	}
+
+	peerRpcPort := viper.GetInt("peerRpc.port")
+	if peerRpcPort == 0 {
+		return nil, errors.Errorf("setting peerRpc.port not found")
+	}
+
+	rpcURL, err := url.Parse(fmt.Sprintf("rpc://%s:%s@%s:%d", peerRpcUser, peerRpcPassword, peerRpcHost, peerRpcPort))
+	if err != nil {
+		return nil, errors.Errorf("failed to parse rpc URL: %v", err)
+	}
+
+	// connect to bitcoin node and get the settings
+	b, err := bitcoin.NewFromURL(rpcURL, false)
+	if err != nil {
+		return nil, fmt.Errorf("error connecting to peer: %v", err)
+	}
+
+	settings, err := b.GetSettings()
+	if err != nil {
+		return nil, fmt.Errorf("error getting settings from peer: %v", err)
+	}
+
+	return &settings, nil
 }
